requirements.txt
fastapi
uvicorn
pymongo
openai
ollama
python-dotenv